+++ staffflow-server/src/main/java/kz/zzhalelov/staffflow/server/dto/departmentDto/DepartmentCreateDto.java
@@ -8,5 +8,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class DepartmentCreateDto {
     String name;
-    Integer managerId;
+    Long managerId;
 }
