+++ staffflow-server/src/main/java/kz/zzhalelov/staffflow/server/dto/departmentDto/DepartmentResponseDto.java
@@ -8,5 +8,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class DepartmentResponseDto {
     String name;
-    Integer managerId;
+    Long managerId;
 }
